chore(server): core IoC #94 - getObjectChildrenFactory
