+++ packages/server/modules/core/repositories/objects.ts
@@ -10,6 +10,7 @@ import {
   GetBatchedStreamObjects,
   GetFormattedObject,
   GetObject,
+  GetObjectChildren,
   GetObjectChildrenStream,
   GetStreamObjects,
   StoreClosuresIfNotFound,
@@ -19,6 +20,7 @@ import {
 } from '@/modules/core/domain/objects/operations'
 import { SpeckleObject } from '@/modules/core/domain/objects/types'
 import { SetOptional } from 'type-fest'
+import { set, toNumber } from 'lodash'
 
 const ObjectChildrenClosure = buildTableHelper('object_children_closure', [
   'parent',
@@ -162,4 +164,88 @@ export const getObjectChildrenStreamFactory =
       )
       .orderBy('objects.id')
     return q.stream({ highWaterMark: 500 })
+  }
+
+export const getObjectChildrenFactory =
+  (deps: { db: Knex }): GetObjectChildren =>
+  async ({ streamId, objectId, limit, depth, select, cursor }) => {
+    limit = toNumber(limit || 0) || 50
+    depth = toNumber(depth || 0) || 1000
+
+    let fullObjectSelect = false
+
+    const q = deps.db.with(
+      'object_children_closure',
+      knex.raw(
+        `SELECT objects.id as parent, d.key as child, d.value as mindepth, ? as "streamId"
+        FROM objects
+        JOIN jsonb_each_text(objects.data->'__closure') d ON true
+        where objects.id = ?`,
+        [streamId, objectId]
+      )
+    )
+
+    if (Array.isArray(select)) {
+      select.forEach((field, index) => {
+        q.select(
+          knex.raw('jsonb_path_query(data, :path) as :name:', {
+            path: '$.' + field,
+            name: '' + index
+          })
+        )
+      })
+    } else {
+      fullObjectSelect = true
+      q.select('data')
+    }
+
+    q.select('id')
+    q.select('createdAt')
+    q.select('speckleType')
+    q.select('totalChildrenCount')
+
+    q.from('object_children_closure')
+
+    q.rightJoin('objects', function () {
+      this.on('objects.streamId', '=', 'object_children_closure.streamId').andOn(
+        'objects.id',
+        '=',
+        'object_children_closure.child'
+      )
+    })
+      .where(
+        knex.raw('object_children_closure."streamId" = ? AND parent = ?', [
+          streamId,
+          objectId
+        ])
+      )
+      .andWhere(knex.raw('object_children_closure.mindepth < ?', [depth]))
+      .andWhere(knex.raw('id > ?', [cursor ? cursor : '0']))
+      .orderBy('objects.id')
+      .limit(limit)
+
+    const rows = await q
+
+    if (rows.length === 0) {
+      return { objects: rows, cursor: null }
+    }
+
+    if (!fullObjectSelect)
+      rows.forEach((o, i, arr) => {
+        const no = {
+          id: o.id,
+          createdAt: o.createdAt,
+          speckleType: o.speckleType,
+          totalChildrenCount: o.totalChildrenCount,
+          data: {}
+        }
+        let k = 0
+        for (const field of select || []) {
+          set(no.data, field, o[k++])
+        }
+        arr[i] = no
+      })
+
+    const lastId = rows[rows.length - 1].id
+    return { objects: rows, cursor: lastId }
   }
