+++ packages/server/modules/core/services/objects.js
@@ -6,91 +6,6 @@ const Objects = () => knex('objects')
 
 module.exports = {
   /**
-   * @returns {Promise<{objects: import('@/modules/core/helpers/types').ObjectRecord[], cursor: string | null}>}
-   */
-  async getObjectChildren({ streamId, objectId, limit, depth, select, cursor }) {
-    limit = parseInt(limit) || 50
-    depth = parseInt(depth) || 1000
-
-    let fullObjectSelect = false
-
-    const q = knex.with(
-      'object_children_closure',
-      knex.raw(
-        `SELECT objects.id as parent, d.key as child, d.value as mindepth, ? as "streamId"
-        FROM objects
-        JOIN jsonb_each_text(objects.data->'__closure') d ON true
-        where objects.id = ?`,
-        [streamId, objectId]
-      )
-    )
-
-    if (Array.isArray(select)) {
-      select.forEach((field, index) => {
-        q.select(
-          knex.raw('jsonb_path_query(data, :path) as :name:', {
-            path: '$.' + field,
-            name: '' + index
-          })
-        )
-      })
-    } else {
-      fullObjectSelect = true
-      q.select('data')
-    }
-
-    q.select('id')
-    q.select('createdAt')
-    q.select('speckleType')
-    q.select('totalChildrenCount')
-
-    q.from('object_children_closure')
-
-    q.rightJoin('objects', function () {
-      this.on('objects.streamId', '=', 'object_children_closure.streamId').andOn(
-        'objects.id',
-        '=',
-        'object_children_closure.child'
-      )
-    })
-      .where(
-        knex.raw('object_children_closure."streamId" = ? AND parent = ?', [
-          streamId,
-          objectId
-        ])
-      )
-      .andWhere(knex.raw('object_children_closure.mindepth < ?', [depth]))
-      .andWhere(knex.raw('id > ?', [cursor ? cursor : '0']))
-      .orderBy('objects.id')
-      .limit(limit)
-
-    const rows = await q
-
-    if (rows.length === 0) {
-      return { objects: rows, cursor: null }
-    }
-
-    if (!fullObjectSelect)
-      rows.forEach((o, i, arr) => {
-        const no = {
-          id: o.id,
-          createdAt: o.createdAt,
-          speckleType: o.speckleType,
-          totalChildrenCount: o.totalChildrenCount,
-          data: {}
-        }
-        let k = 0
-        for (const field of select) {
-          set(no.data, field, o[k++])
-        }
-        arr[i] = no
-      })
-
-    const lastId = rows[rows.length - 1].id
-    return { objects: rows, cursor: lastId }
-  },
-
-  /**
    *
    * This query is inefficient on larger sets (n * 10k objects) as we need to return the total count on an arbitrarily (user) defined selection of objects.
    * A possible future optimisation route would be to cache the total count of a query (as objects are immutable, it will not change) on a first run, and, if found on a subsequent round, do a simpler query and merge the total count result.
